Phone: Rename `ui` package to `presentation`
